backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openslide-bin==4.0.0.8
openslide-python==1.4.2
pillow==11.3.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
zipp==3.23.0
